pyproject.toml
[project]
name = "1500-q-a"
version = "0.1.0"
description = ""
authors = [
    {name = "Laura Zhu",email = "148168410+zarual@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.82.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "regex (>=2024.11.6,<2025.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "pydub (>=0.25.1)",
    "websockets (>=11.0,<12.0)",
    "numpy (>=2.2.6,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
